\name{Intelligence data}
\alias{IntelligenceData}
\docType{data}
\title{Intelligence data (number completion)}
\description{
The data are dichotomous scores of a Dutch intelligence test on number completion (Dutch: "Cijferreeksen", Drenth and Hoolwerf, 1970).  The file consists of archival data that were collected in a high-stakes personnel selection context around 1990. 
}
\usage{data(IntelligenceData)}
\format{A 1000x26 matrix of dichotomous item scores.}
\references{
Drenth, P. J. D., and Hoolwerf, G. (1970) \emph{Numerieke aanleg test - Cijferreeksen (NAT-Cijferreeksen)} [Numerical ability test]. Amsterdam: The Netherlands.
}
\examples{data(IntelligenceData)}
\keyword{datasets}
