\name{Physical Functioning}
\alias{PhysFuncData}
\docType{data}
\title{The SF-36 physical functioning data}
\description{
These data are from the Physical Functioning scale of the SF-36 (Ware and Sherbourne, 1992). Data consist of scores of 714 respondents on 10 polytomously scored items (0 = no, not limited at all; 1 = limited a little; 2 = limited a lot). 
}
\usage{data(PhysFuncData)}
\format{A 714x10 matrix of polytomous item scores (scores 0, 1, and 2).}
\references{
Ware, J. E., Jr., and Sherbourne, C. D. (1992) The MOS 36-item short-form health survey (SF-36): Conceptual framework and item selection. \emph{Medical Care}, \bold{30}, 473--483.
}
\examples{data(PhysFuncData)}
\keyword{datasets}
